Working serial sender on the real hardware!
This is a barebones test for the rom1 cpu that just
sends "-" to the serial port over and over again.

The SIO hardware takes its clock from the CTC. So
you need to initialize the CTC and SIO for the communication.
The used settings create a 38400baus 8 databit 1,5 stop bit, odd
parity link.